requirements.txt
# =============================================================================
# CogniShare Protocol - Dependencies
# =============================================================================
# A decentralized RAG system with x402 micropayments on Cronos EVM
# Hackathon MVP - December 2024
# =============================================================================

# ----- Web UI -----
streamlit>=1.28.0
plotly>=5.18.0

# ----- AI/LLM Stack -----
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.6.0

# ----- Vector Database (NAMEN GEÄNDERT!) -----
pinecone>=3.0.0

# ----- Blockchain / Web3 -----
web3>=6.11.0

# ----- Document Processing -----
pypdf2>=3.0.0

# ----- Data Analysis -----
pandas>=2.0.0

# ----- Utilities -----
python-dotenv>=1.0.0
tiktoken>=0.5.0

# ----- Smart Contract Deployment -----
py-solc-x>=2.0.0

# ----- Market Data (Crypto Prices) -----
pycoingecko>=3.1.0